requirements.txt
amqp==5.0.2
asgiref==3.2.10
billiard==3.6.3.0
boto3==1.15.7
botocore==1.18.7
celery==5.0.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
code128==0.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.1
django-3-jet==1.0.8
django-cors-headers==3.5.0
django-import-export==2.4.0
django-jet==1.0.8
django-storages==1.10.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.7.0
drf-yasg==1.17.1
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
jet==1.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.4
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
python-barcode==0.13.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-postmark==0.5.5
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0


